refactor: view폴더 내 파일 console, web으로 폴더 분리
