+++ src/view/web/WebInput.js
@@ -1,5 +1,5 @@
-import { LOTTO_RULE } from "../constant/rule.js";
-import Validator from "../Validator/validator.js";
+import { LOTTO_RULE } from "../../constant/rule.js";
+import Validator from "../../Validator/validator.js";
 import WebInputHandler from "./WebInputHandler.js";
 
 const WebInput = {
